backend/requirements.txt
Django>=5.0.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
markdown>=3.5.0
django-unfold>=0.20.0 